Renamed leave_application_screen.dart
